backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
httpx>=0.26.0
chromadb>=0.4.22
sentence-transformers>=2.4.0
torch>=2.2.0
torchvision>=0.17.0
faiss-cpu>=1.8.0
Pillow>=10.2.0
pydantic>=2.6.0
